fix ubsan integer implicit conversion
see https://cirrus-ci.com/task/5765846023798784

+++ src/netbase.cpp
@@ -209,7 +209,7 @@ bool Lookup(const std::string& name, std::vector<CService>& vAddr, int portDefau
         return false;
     vAddr.resize(vIP.size());
     for (unsigned int i = 0; i < vIP.size(); i++)
-        vAddr[i] = CService(vIP[i], port);
+        vAddr[i] = CService(vIP[i], (uint16_t)port);
     return true;
 }
